use more int64_t

+++ sperl_array.c
@@ -6,7 +6,7 @@
 #include "sperl_allocator_util.h"
 #include "sperl_parser.h"
 
-SPerl_ARRAY* SPerl_ARRAY_new(SPerl* sperl, size_t capacity) {
+SPerl_ARRAY* SPerl_ARRAY_new(SPerl* sperl, int64_t capacity) {
   
   SPerl_ARRAY* array = SPerl_ALLOCATOR_UTIL_safe_malloc(1, sizeof(SPerl_ARRAY));
   array->length = 0;
@@ -27,13 +27,13 @@ SPerl_ARRAY* SPerl_ARRAY_new(SPerl* sperl, size_t capacity) {
 void SPerl_ARRAY_push(SPerl* sperl, SPerl_ARRAY* array, const void* value) {
 
   size_t length = array->length;
-  size_t capacity = array->capacity;
+  int64_t capacity = array->capacity;
   
   if (length >= capacity) {
     if (capacity > SIZE_MAX / 2) {
       SPerl_ALLOCATOR_UTIL_exit_with_malloc_failure();
     }
-    size_t new_capacity = capacity * 2;
+    int64_t new_capacity = capacity * 2;
     array->values = (void**) SPerl_ALLOCATOR_UTIL_safe_realloc(array->values, new_capacity, sizeof(void*));
     array->capacity = new_capacity;
   }
