+++ sperl_array.h
@@ -7,11 +7,11 @@
 struct SPerl_array {
   void** values;
   size_t length;
-  size_t capacity;
+  int64_t capacity;
 };
 
 // Array function
-SPerl_ARRAY* SPerl_ARRAY_new(SPerl* sperl, size_t capacity);
+SPerl_ARRAY* SPerl_ARRAY_new(SPerl* sperl, int64_t capacity);
 void SPerl_ARRAY_push(SPerl* sperl, SPerl_ARRAY* array, const void* value);
 void* SPerl_ARRAY_fetch(SPerl* sperl, SPerl_ARRAY* array, int64_t index);
 void* SPerl_ARRAY_pop(SPerl* sperl, SPerl_ARRAY* array);
